Split vector<>sets from vector transforms
(and add a few functions ;)
